refactor: Price 클래스로 할인 계산 기능 이동
