+++ domain/src/main/java/com/example/domain/model/price/Price.kt
@@ -1,9 +1,26 @@
 package com.example.domain.model.price
 
-@JvmInline
-value class Price(val price: Int = DEFAULT) {
+import com.example.domain.model.policy.MorningPolicy
+import com.example.domain.model.policy.MovieDayPolicy
+import com.example.domain.model.policy.NightPolicy
+import java.time.LocalDate
+import java.time.LocalTime
+
+data class Price(val price: Int = DEFAULT) {
+    private val policies = listOf(MovieDayPolicy(), MorningPolicy(), NightPolicy())
+
     init {
         require(price >= 0) { throw IllegalArgumentException(MINUS_ERROR) }
+    }
+
+    fun calculate(
+        playingDate: LocalDate,
+        playingTime: LocalTime
+    ): Price {
+        val calculatePrice = policies.fold(Price(price)) { calculatePrice, policy ->
+            policy.calculate(playingDate, playingTime, calculatePrice)
+        }
+        return calculatePrice
     }
 
     companion object {
