+++ domain/src/test/java/PriceCalculatorTest.kt
@@ -1,5 +1,4 @@
 import com.example.domain.model.price.Price
-import com.example.domain.model.price.PriceCalculator
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
 import java.time.LocalDate
@@ -8,8 +7,7 @@ import java.time.LocalTime
 class PriceCalculatorTest {
     @Test
     fun `무비데이와 조조가 겹치면 무비데이 할인이 선적용된 가격이 계산된다`() {
-        val actual =
-            PriceCalculator.calculate(Price(), LocalDate.of(2023, 3, 10), LocalTime.of(10, 0))
+        val actual = Price().calculate(LocalDate.of(2023, 3, 10), LocalTime.of(10, 0))
         val expected = Price(9700)
         assertEquals(actual, expected)
     }
